+++ src/observable/observable_block_all.rs
@@ -46,9 +46,6 @@ where
     (self.complete)();
     self.is_stopped.store(true, Ordering::Relaxed);
   }
-
-  #[inline]
-  fn is_stopped(&self) -> bool { self.is_stopped.load(Ordering::Relaxed) }
 }
 
 pub trait SubscribeBlockingAll<'a, N, E, C> {