+++ src/observable/observable_comp.rs
@@ -4,7 +4,6 @@ use crate::prelude::*;
 pub struct ObserverComp<N, C, Item> {
   next: N,
   complete: C,
-  is_stopped: bool,
   _marker: TypeHint<*const Item>,
 }
 
@@ -18,12 +17,8 @@ where
   #[inline]
   fn next(&mut self, value: Item) { (self.next)(value); }
   #[inline]
-  fn error(&mut self, _err: ()) { self.is_stopped = true; }
-  fn complete(&mut self) {
-    (self.complete)();
-    self.is_stopped = true;
-  }
-  fn is_stopped(&self) -> bool { self.is_stopped }
+  fn error(&mut self, _err: ()) {}
+  fn complete(&mut self) { (self.complete)(); }
 }
 
 pub trait SubscribeComplete<'a, N, C> {
@@ -59,7 +54,6 @@ where
     let unsub = self.actual_subscribe(Subscriber::local(ObserverComp {
       next,
       complete,
-      is_stopped: false,
       _marker: TypeHint::new(),
     }));
     SubscriptionWrapper(unsub)
@@ -85,7 +79,6 @@ where
     let unsub = self.0.actual_subscribe(Subscriber::shared(ObserverComp {
       next,
       complete,
-      is_stopped: false,
       _marker: TypeHint::new(),
     }));
     SubscriptionWrapper(unsub)