+++ src/ops/group_by.rs
@@ -1,5 +1,5 @@
 use crate::prelude::*;
-use crate::{complete_proxy_impl, error_proxy_impl, is_stopped_proxy_impl};
+use crate::{complete_proxy_impl, error_proxy_impl};
 
 use std::{clone::Clone, cmp::Eq, collections::HashSet, hash::Hash};
 
@@ -26,7 +26,6 @@ where
   }
   error_proxy_impl!(Err, observer);
   complete_proxy_impl!(observer);
-  is_stopped_proxy_impl!(observer);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -136,7 +135,6 @@ where
   }
   error_proxy_impl!(Err, observer);
   complete_proxy_impl!(observer);
-  is_stopped_proxy_impl!(observer);
 }
 
 ///////////////////////////////////////////////////////////////////////////////