+++ src/ops/skip_last.rs
@@ -1,5 +1,5 @@
 use crate::prelude::*;
-use crate::{complete_proxy_impl, error_proxy_impl, is_stopped_proxy_impl};
+use crate::{complete_proxy_impl, error_proxy_impl};
 use std::collections::VecDeque;
 
 #[derive(Clone)]
@@ -71,7 +71,6 @@ where
 
   error_proxy_impl!(Err, observer);
   complete_proxy_impl!(observer);
-  is_stopped_proxy_impl!(observer);
 }
 
 #[cfg(test)]