+++ src/ops/throttle_time.rs
@@ -170,11 +170,6 @@ where
     }
     inner.observer.complete();
   }
-
-  fn is_stopped(&self) -> bool {
-    let inner = self.0.lock().unwrap();
-    inner.observer.is_stopped()
-  }
 }
 
 impl<O, S> Observer for LocalThrottleObserver<O, S, O::Item>
@@ -226,11 +221,6 @@ where
       inner.observer.next(value);
     }
     inner.observer.complete();
-  }
-
-  fn is_stopped(&self) -> bool {
-    let inner = self.0.borrow_mut();
-    inner.observer.is_stopped()
   }
 }